Reorganize Files into More Extensible Structure
File Upload Module: Allow users to upload CRM extracts.
Column Matching Module: Interface to call OpenAPI for column matching and allow user edits.
Criteria Selection Module: GUI for users to select checkboxes that define a 'complete record'.
Data Visualization Module: Display percentage of filled values and overall 'complete record' percentage.
Data Enrichment Module: Provide options for email validation, phone number validation, and third-party data enrichment.
Download Module: Allow users to download the enriched and validated data and reports.
