--- app/email_manager.py
@@ -1,9 +1,0 @@
-# app/email_manager.py
-#import streamlit as st
-#from .report import generate_report
-
-
-def capture_user_email(user_email=None, complete_records=None, incomplete_records=None):
-    # Capture user email for marketing database as well as send report
-    # Put email into DB & Gather report materials
-    pass
