+++ app/services/api_service.py
@@ -1,0 +1,3 @@
+def match_columns(file):
+    # Code to call OpenAPI and match columns
+    pass
