+++ app/services/boilerplate.py
@@ -1,5 +1,7 @@
-#app/report/boilerplate.py
+#frontend/report/boilerplate.py
 # This file holds the boilerplate strings
+
+
 
 headers_how_to_use_this_tool ="""
                     **Purpose:**
