+++ app/services/data_enrichment_service.py
@@ -1,0 +1,3 @@
+def enrich_data(data):
+    # Code to enrich data
+    pass
