+++ app/services/data_model_validation_service.py
@@ -1,0 +1,3 @@
+def validate_and_transform_data(data, model):
+    # Code to validate and transform data
+    pass
