+++ app/services/email_validation.py
@@ -1,0 +1,3 @@
+def validate_email(email):
+    # Code to validate email
+    pass
