+++ app/services/phone_validation.py
@@ -1,0 +1,3 @@
+def validate_phone(phone):
+    # Code to validate phone number
+    pass
