+++ app/utils/data_transformations.py
@@ -1,0 +1,3 @@
+def transform_data(data, transformations):
+    # Code to apply transformations
+    pass
