+++ app/utils/validation_utils.py
@@ -1,0 +1,7 @@
+def is_valid_email(email):
+    # Code to check if email is valid
+    pass
+
+def is_valid_phone(phone):
+    # Code to check if phone is valid
+    pass
